+++ packages/thumbnail/src/ThumbnailListWithStore.tsx
@@ -14,12 +14,14 @@ import { SpinnerContext } from './SpinnerContext';
 import { ThumbnailList } from './ThumbnailList';
 import { usePageLabels } from './usePageLabels';
 import type { RenderCurrentPageLabel } from './types/RenderCurrentPageLabelProps';
+import type { RenderThumbnail } from './types/RenderThumbnailProps';
 import type { StoreProps } from './types/StoreProps';
 
 export const ThumbnailListWithStore: React.FC<{
     renderCurrentPageLabel?: RenderCurrentPageLabel;
+    renderThumbnail?: RenderThumbnail;
     store: Store<StoreProps>;
-}> = ({ renderCurrentPageLabel, store }) => {
+}> = ({ renderCurrentPageLabel, renderThumbnail, store }) => {
     const labels = usePageLabels(store);
     const [currentDoc, setCurrentDoc] = React.useState<PdfJs.PdfDocument>();
     const [currentPage, setCurrentPage] = React.useState(store.get('currentPage') || 0);
@@ -84,6 +86,7 @@ export const ThumbnailListWithStore: React.FC<{
             pageHeight={pageHeight}
             pageWidth={pageWidth}
             renderCurrentPageLabel={renderCurrentPageLabel}
+            renderThumbnail={renderThumbnail}
             rotation={rotation}
             onJumpToPage={jump}
         />