+++ packages/thumbnail/src/types/RenderThumbnailProps.ts
@@ -1,0 +1,20 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2021 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+import * as React from 'react';
+
+export interface RenderThumbnailProps {
+    currentPage: number;
+    numPages: number;
+    pageIndex: number;
+    renderPageLabel: React.ReactElement;
+    renderPageThumbnail: React.ReactElement;
+    onJumpToPage: () => void;
+}
+
+export type RenderThumbnail = (props: RenderThumbnailProps) => React.ReactElement;